* commit '3c58da56299fdb3b5a1ab39ed895d97bda18495e':
  treewide: use std::min/max in place of open-coding it
  freebusy: use size_t for RecurrenceState::ParseBlob
  freebusy: use size_t for RecurrenceState::GetBlob
  common: use size_t in ECChannel size parameters
  daemons: make use of ec_log_always

+++ libfreebusy/recurrence.h
@@ -21,7 +21,7 @@ namespace KC {
 class _kc_export recurrence _kc_final {
 public:
 	HRESULT HrLoadRecurrenceState(const char *data, size_t len, ULONG flags) { return m_sRecState.ParseBlob(data, len, flags); }
-	HRESULT HrGetRecurrenceState(char **lppData, unsigned int *lpulLen, void *base = NULL);
+	HRESULT HrGetRecurrenceState(char **lppData, size_t *lpulLen, void *base = NULL);
 	void HrGetHumanReadableString(std::string *);
 	HRESULT HrGetItems(time_t start, time_t end, const TIMEZONE_STRUCT &ttZinfo, ULONG ulBusyStatus, OccrInfo **lppFbBlock, ULONG *lpcValues, bool last = false);
 	enum freq_type { DAILY, WEEKLY, MONTHLY, YEARLY };
